Add SetMaxLines and move the method to top bar
I added the setmaxlines in the json formater to stop endless scrolling in the json
view window. This stops it from endless scrolling but as soon as you hit the end it
snaps to the top of the text document. I will have to see how I can fix this.

I moved the method to the top bar. This will be used the most and I feel it should be
at the top as it would be something that would be changed often. I will be adding pages
to the system and having to switch pages to change this option would be a pain.